requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
api==0.0.7
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==5.0.1
bqplot==0.11.0
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cli-helpers==2.3.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
configobj==5.0.8
contourpy==1.0.6
cryptography==36.0.2
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.3
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.5
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
findspark==2.0.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.3.3
Folio==0.4
folium==0.14.0
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.4
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.4.0
ghp-import==2.1.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpt==0.2
gpt3==0.0.1
greenlet==2.0.2
grpcio==1.54.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.8.0
html2text==2020.1.16
html5lib==1.1
idna==3.4
imageio==2.28.0
ipykernel==6.23.1
ipympl==0.9.3
ipython==8.13.2
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==7.7.5
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.2
jellyfish==0.11.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.1.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
jupyterlab_server==2.22.1
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
Levenshtein==0.21.0
libclang==16.0.0
llvmlite==0.41.0
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
mycli==1.27.0
mysql-connector-python==8.0.33
nb==0.1.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.4.0
nbformat==5.8.0
nbmanips==2.0.1
nest-asyncio==1.5.6
networkx==3.1
nlp==0.4.0
nodejs==0.1.1
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optional-django==0.1.0
outcome==1.2.0
packaging==23.1
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
pip-review==1.3.0
platformdirs==3.5.1
plotly==5.14.1
prettytable==3.8.0
prometheus-client==0.15.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
py4j==0.10.9.5
pyaes==1.6.1
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyMySQL==1.1.0
pyodbc==4.0.35
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
pyspark==3.3.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
python-Levenshtein==0.21.0
pytz==2022.7
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.0
QtPy==2.3.0
rapidfuzz==3.0.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.9.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sql==2022.4.0
SQLAlchemy==2.0.13
sqlglot==18.11.4
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
sympy==1.11.1
tabulate==0.9.0
tb-nightly==2.13.0a20230426
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.13
voila==0.4.0
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==2.2.2
widgetsnbextension==3.6.4
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.6
xlrd==2.0.1
xxhash==3.2.0
yapf==0.33.0
yarl==1.8.2
yfinance==0.2.3
